- Working on making the MobileSelectOptions component usable.
Created another util file called categoriesActionEvents. Transferred the action events of the custom selects inside the CategorySelect inside the new categoriesActionEvents. Reason for this is for reusing the functions and so that the CategorySelect component wont be cluttered with event functions.

Transfered the Feature, PlaceList, LocationInfoInput types to the scheduleType file. This is for reusing purposes.

Adjusted the CustomSelect reusable component. Make it so that it will display the list of choices based on what devices you are on. If you are in laptop or above screens it will display the normal select and if you are in mobile it will display a popup list this is the purpose of MobileSelectOptions component.

Updated the MobileSelectOptions component.