- Restructuring CategorySelect component.
Created another util called TanStackQueryFns. This utils will store all of the reusable tanstack queries.

Deleted a util called getLocationInfoinitials, replace it with a util called stateInitials. This will store all of the state initials.

Adjusted the names of some of the props in MobileSelectOptions component.

Transferred the query functions in CategorySelect and transfer it into TanStackQuery fns.

- Note: Will be restructuring it more tomorrow.