- Working on DesktopSelectOption component.
Adjusted the props types and an onclick function of the CustomSelect component (FIle: CustomSelect.tsx)

Adjusted the props that has been passed to the CustomSelect component. (File: CategorySelect.tsx)

Remove the map function children of CustomSelect and replace it with the DesktopSelectOptions component and passed the necessary props. (File: CategorySelect.tsx)