- Updating the ScheduleForm, and CalendarNav components.
Added a util function called formatSelectedDate in useDate utils (file: useDate.ts)

Updated the CalendarNav component. Adjusted the setSelectedDate function, instead of setting the local state it is now setting the global state in the useDateStore. It is also using  the dateSelected from the useDateStore instead of using the local state (file: CalendarNav.tsx)

Adjusted the ScheduleForm onSuccess. It now changes the selectedDate in the CalendarNav so it will now properly display the new added Schedule. (file: ScheduleForm.tsx)

- CategorySelect quick fixes

Fix the setOptionType and the setMobileOptionHeader in the setToggleMobileOptions props of the CustomSelect. Changed it to its respective value instead of using the same value across the thre CustomSelect.