- Minor adjustments on the schedule page.
Adjusted the schedule page. Instead of passing an array to the Schedules component it now passes an object. (File: Schedules.tsx)

Minor adjustment on the Schedule component. Instead of displaying the current date using a state, it now uses the Date object directly that has a parameter of dateSelected from the useDateStore. (File: Schedule.tsx)

Adjusted the CalendarNav and some minor CSS adjustments. Remove the two custom select, replaces it with just a text showing the month and the current/selected date. (File: CalendarNav.tsx)