- Working on making the supabase work with fetch/API routs. Minor changes for the middleware. Supabase Utils. Minor changes to the authActions and todolist action.
-- API Routes changes
Added another api route called getTodoList.

Adjusted the todolist page to be using a fetch instead of calling the get directly. This is to utilize the caching of nextjs. Will be applying it to the rest of the pages. (File: page.tsx of TodoList)

Minor adjustments for the getTodoLists util function. Adjusted the return value to be the same as the getTodoList api route. (File: getTodoLists.ts)

Adjusted the todolist component because of the new api route and the new return value of the getTodoList util. (File: TodoLists.tsx)

-- Supabase Utils
Added a supabase utils to be reused across the api routes/utils/pages etc.. (File: supabaseUtils.ts)

-- Middleware changes

Added a new utils in cookiesUtils called supabaseAuth. This uses the NexRequest to get the cookie header. (File: middleware.ts)

Adjusted the middleware to use the supabaseAuth instead of the getCookieAuth. (File: middleware.ts)

-- TodoList and auth action changes

Adjusted the signOut action in the authAction. (File: authActions.ts)

Adjusted the mutateTodoList action in todolistActions. Added a revalidate tag to revalidate the todolist cache. (File: todolistActions.ts)